Rename plugins to Actinos.
