+++ hack/.golint_failures
@@ -21,7 +21,7 @@ pkg/batchd/client/informers/internalinterfaces
 pkg/batchd/controller/policy
 pkg/batchd/controller/queue
 pkg/batchd/policy
-pkg/batchd/policy/plugins/drf
+pkg/batchd/policy/actions/allocate
 pkg/batchd/policy/util
 pkg/queuejob-ctrl/apis/v1
 pkg/queuejob-ctrl/client
