+++ pkg/batchd/policy/actions/allocate/podset_info.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package drf
+package allocate
 
 import (
 	"k8s.io/api/core/v1"
